+++ src/cli/podmander-podctl.ads
@@ -1,0 +1,5 @@
+--  Copyright (C) 2026 Jochen Lillich
+--  SPDX-License-Identifier: Apache-2.0
+
+package Podmander.Podctl is
+end Podmander.Podctl;