+++ tests/podmander-podctl-config_tests.ads
@@ -1,0 +1,10 @@
+--  Copyright (C) 2026 Jochen Lillich
+--  SPDX-License-Identifier: Apache-2.0
+
+with AUnit.Test_Suites;
+
+package Podmander.Podctl.Config_Tests is
+
+   function Suite return AUnit.Test_Suites.Access_Test_Suite;
+
+end Podmander.Podctl.Config_Tests;