+++ tests/test_runner.adb
@@ -19,27 +19,39 @@ with Podmander.Database_Tests;
 with Podmander.Generators.Quadlet_Tests;
 with Podmander.Logging_Tests;
 with Podmander.Messages_Tests;
+with Podmander.Podctl.Config_Tests;
 
 procedure Test_Runner is
 
    function All_Suites return AUnit.Test_Suites.Access_Test_Suite is
-      Result : constant AUnit.Test_Suites.Access_Test_Suite := new AUnit.Test_Suites.Test_Suite;
+      Result : constant AUnit.Test_Suites.Access_Test_Suite :=
+        new AUnit.Test_Suites.Test_Suite;
    begin
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Agent.Host_Command_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Agent.Host_Command.Result_Mapping_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Agent.Host_Command_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Agent.Host_Command.Result_Mapping_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Messages_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Agent.Repository_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Registrar_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Scheduler_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Service.Repository_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Service_Catalog.Repository_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Controller.Stack_Submission_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Agent.Repository_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Registrar_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Scheduler_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Service.Repository_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Service_Catalog.Repository_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Controller.Stack_Submission_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Controller_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Enrollment_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Logging_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Config_Tests.Suite);
-      AUnit.Test_Suites.Add_Test (Result, Podmander.Generators.Quadlet_Tests.Suite);
+      AUnit.Test_Suites.Add_Test
+        (Result, Podmander.Generators.Quadlet_Tests.Suite);
       AUnit.Test_Suites.Add_Test (Result, Podmander.Database_Tests.Suite);
+      AUnit.Test_Suites.Add_Test (Result, Podmander.Podctl.Config_Tests.Suite);
       return Result;
    end All_Suites;